backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and ML
openai==1.3.7
pinecone-client==2.2.4
langchain==0.0.350
langchain-openai==0.0.2

# Data processing
pandas==2.1.3
numpy==1.25.2
python-docx==1.1.0
PyPDF2==3.0.1

# HTTP and scraping
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0 